Extract the HTML template into its own file
