--- config/app-info.js
@@ -1,26 +1,0 @@
-const url = "https://kristorres.github.io";
-
-export default {
-    name: "Kris Torres’s Website",
-    description: "Kris Torres’s website.",
-    author: "Kris Torres",
-    url,
-    keywords: [
-        "Kris Torres",
-        "computer programming",
-        "computer programmer",
-        "developer",
-        "website",
-        "portfolio"
-    ],
-    twitter: {
-        handle: "@ohayoukris",
-        bannerURL: `${url}/images/banners/kt-banner-main.png`,
-        description: [
-            "Hello, world!",
-            "I’m an app developer and I love to build cool things! 👨🏻‍💻🗼😄"
-        ].join(" ")
-    },
-    themeColor: "#1A64D7",
-    iconURL: "/images/logos/KT_PRI_LOGO_ON_BLU.png"
-};
